backend/requirements.txt
asgiref==3.11.1
boto3==1.42.68
botocore==1.42.68
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
django-anymail==14.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-simple-history==3.11.0
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==40.11.0
idna==3.11
jmespath==1.1.0
packaging==26.0
pipdeptree==2.31.0
psycopg2==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
requests==2.32.5
ruff==0.15.6
s3transfer==0.16.0
sentry-sdk==2.54.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
